// apps/recommendations-service/src/__generated__/resolvers-types.d.ts
export type Maybe<T> = T | null;
// ...
export type RequireFields<T, K extends keyof T> = Omit<T, K> & {
    [P in K]-?: NonNullable<T[P]>;
};
export type Scalars = {
    ID: {
        input: string;
        output: string;
    };
    String: {
        input: string;
        output: string;
    };
    Boolean: {
        input: boolean;
        output: boolean;
    };
    Int: {
        input: number;
        output: number;
    };
    Float: {
        input: number;
        output: number;
    };
    Date: {
        input: any;
        output: any;
    };
    DateTime: {
        input: any;
        output: any;
    };
    _FieldSet: {
        input: any;
        output: any;
    };
};
export type Case = {
    __typename?: 'Case';
    id: Scalars['ID']['output'];
    recommendations: Array<Recommendation>;
};
export type CreateRecommendationInput = {
    caseId: Scalars['ID']['input'];
    description: Scalars['String']['input'];
    priority: Priority;
    providerId: Scalars['ID']['input'];
    title: Scalars['String']['input'];
};
export type Mutation = {
    __typename?: 'Mutation';
    createRecommendation?: Maybe<Recommendation>;
    updateRecommendationStatus?: Maybe<Recommendation>;
};
export type MutationCreateRecommendationArgs = {
    input: CreateRecommendationInput;
};
export type MutationUpdateRecommendationStatusArgs = {
    id: Scalars['ID']['input'];
    status: RecommendationStatus;
};
export declare enum Priority {
    High = "HIGH",
    Low = "LOW",
    Medium = "MEDIUM",
    Urgent = "URGENT"
}
export type Provider = {
    __typename?: 'Provider';
    id: Scalars['ID']['output'];
    recommendations: Array<Recommendation>;
};
export type Query = {
    __typename?: 'Query';
    recommendation?: Maybe<Recommendation>;
    recommendationsByProvider: Array<Recommendation>;
    recommendationsForCase: Array<Recommendation>;
};
export type QueryRecommendationArgs = {
    id: Scalars['ID']['input'];
};
export type QueryRecommendationsByProviderArgs = {
    providerId: Scalars['ID']['input'];
};
export type QueryRecommendationsForCaseArgs = {
    caseId: Scalars['ID']['input'];
};
export type Recommendation = {
    __typename?: 'Recommendation';
    caseId: Scalars['ID']['output'];
    createdAt: Scalars['DateTime']['output'];
    description: Scalars['String']['output'];
    id: Scalars['ID']['output'];
    priority: Priority;
    providerId: Scalars['ID']['output'];
    status: RecommendationStatus;
    title: Scalars['String']['output'];
    updatedAt: Scalars['DateTime']['output'];
};
export type RecommendationItem = {
    __typename?: 'RecommendationItem';
    id: Scalars['ID']['output'];
};
export declare enum RecommendationStatus {
    Active = "ACTIVE",
    Cancelled = "CANCELLED",
    Completed = "COMPLETED",
    Draft = "DRAFT"
}
export type WithIndex<TObject> = TObject & Record<string, any>;
export type ResolversObject<TObject> = WithIndex<TObject>;
export type ResolverTypeWrapper<T> = Promise<T> | T;
export type ReferenceResolver<TResult, TReference, TContext> = (reference: TReference, context: TContext, info: GraphQLResolveInfo) => Promise<TResult> | TResult;
type ScalarCheck<T, S> = S extends true ? T : NullableCheck<T, S>;
type NullableCheck<T, S> = Maybe<T> extends T ? Maybe<ListCheck<NonNullable<T>, S>> : ListCheck<T, S>;
type ListCheck<T, S> = T extends (infer U)[] ? NullableCheck<U, S>[] : GraphQLRecursivePick<T, S>;
export type GraphQLRecursivePick<T, S> = {
    [K in keyof T & keyof S]: ScalarCheck<T[K], S[K]>;
};
export type ResolverWithResolve<TResult, TParent, TContext, TArgs> = {
    resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | ResolverWithResolve<TResult, TParent, TContext, TArgs>;
// ...
export type IsTypeOfResolverFn<T = {}, TContext = {}> = (obj: T, context: TContext, info: GraphQLResolveInfo) => boolean | Promise<boolean>;
export type NextResolverFn<T> = () => Promise<T>;
export type DirectiveResolverFn<TResult = {}, TParent = {}, TContext = {}, TArgs = {}> = (next: NextResolverFn<TResult>, parent: TParent, args: TArgs, context: TContext, info: GraphQLResolveInfo) => TResult | Promise<TResult>;
export type ResolversTypes = ResolversObject<{
    Case: ResolverTypeWrapper<Case>;
    ID: ResolverTypeWrapper<Scalars['ID']['output']>;
    CreateRecommendationInput: CreateRecommendationInput;
    String: ResolverTypeWrapper<Scalars['String']['output']>;
    Date: ResolverTypeWrapper<Scalars['Date']['output']>;
    DateTime: ResolverTypeWrapper<Scalars['DateTime']['output']>;
    Mutation: ResolverTypeWrapper<{}>;
    Priority: Priority;
    Provider: ResolverTypeWrapper<Provider>;
    Query: ResolverTypeWrapper<{}>;
    Recommendation: ResolverTypeWrapper<Recommendation>;
    RecommendationItem: ResolverTypeWrapper<RecommendationItem>;
    RecommendationStatus: RecommendationStatus;
    Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
}>;
export type ResolversParentTypes = ResolversObject<{
    Case: Case;
    ID: Scalars['ID']['output'];
    CreateRecommendationInput: CreateRecommendationInput;
    String: Scalars['String']['output'];
    Date: Scalars['Date']['output'];
    DateTime: Scalars['DateTime']['output'];
    Mutation: {};
    Provider: Provider;
    Query: {};
    Recommendation: Recommendation;
    RecommendationItem: RecommendationItem;
    Boolean: Scalars['Boolean']['output'];
}>;
export type CaseResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Case'] = ResolversParentTypes['Case']> = ResolversObject<{
    __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['Case']>, {
        __typename: 'Case';
    } & GraphQLRecursivePick<ParentType, {
        "id": true;
    }>, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    recommendations?: Resolver<Array<ResolversTypes['Recommendation']>, ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export interface DateScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['Date'], any> {
    name: 'Date';
}
export interface DateTimeScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['DateTime'], any> {
    name: 'DateTime';
}
export type MutationResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = ResolversObject<{
    createRecommendation?: Resolver<Maybe<ResolversTypes['Recommendation']>, ParentType, ContextType, RequireFields<MutationCreateRecommendationArgs, 'input'>>;
    updateRecommendationStatus?: Resolver<Maybe<ResolversTypes['Recommendation']>, ParentType, ContextType, RequireFields<MutationUpdateRecommendationStatusArgs, 'id' | 'status'>>;
}>;
export type ProviderResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Provider'] = ResolversParentTypes['Provider']> = ResolversObject<{
    __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['Provider']>, {
        __typename: 'Provider';
    } & GraphQLRecursivePick<ParentType, {
        "id": true;
    }>, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    recommendations?: Resolver<Array<ResolversTypes['Recommendation']>, ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export type QueryResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = ResolversObject<{
    recommendation?: Resolver<Maybe<ResolversTypes['Recommendation']>, ParentType, ContextType, RequireFields<QueryRecommendationArgs, 'id'>>;
    recommendationsByProvider?: Resolver<Array<ResolversTypes['Recommendation']>, ParentType, ContextType, RequireFields<QueryRecommendationsByProviderArgs, 'providerId'>>;
    recommendationsForCase?: Resolver<Array<ResolversTypes['Recommendation']>, ParentType, ContextType, RequireFields<QueryRecommendationsForCaseArgs, 'caseId'>>;
}>;
export type RecommendationResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['Recommendation'] = ResolversParentTypes['Recommendation']> = ResolversObject<{
    __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['Recommendation']>, {
        __typename: 'Recommendation';
    } & GraphQLRecursivePick<ParentType, {
        "id": true;
    }>, ContextType>;
    caseId?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    createdAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
    description?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    priority?: Resolver<ResolversTypes['Priority'], ParentType, ContextType>;
    providerId?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    status?: Resolver<ResolversTypes['RecommendationStatus'], ParentType, ContextType>;
    title?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
    updatedAt?: Resolver<ResolversTypes['DateTime'], ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export type RecommendationItemResolvers<ContextType = DataSourceContext, ParentType extends ResolversParentTypes['RecommendationItem'] = ResolversParentTypes['RecommendationItem']> = ResolversObject<{
    __resolveReference?: ReferenceResolver<Maybe<ResolversTypes['RecommendationItem']>, ParentType, ContextType>;
    id?: Resolver<ResolversTypes['ID'], ParentType, ContextType>;
    __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
export type Resolvers<ContextType = DataSourceContext> = ResolversObject<{
    Case?: CaseResolvers<ContextType>;
    Date?: GraphQLScalarType;
    DateTime?: GraphQLScalarType;
    Mutation?: MutationResolvers<ContextType>;
    Provider?: ProviderResolvers<ContextType>;
    Query?: QueryResolvers<ContextType>;
    Recommendation?: RecommendationResolvers<ContextType>;
    RecommendationItem?: RecommendationItemResolvers<ContextType>;
}>;